requirements.txt
# Core Dependencies
pyautogui>=0.9.50
pillow>=10.4.0
pytesseract>=0.3.10
opencv-python>=4.8.0

# Database & Storage
SQLAlchemy>=2.0.20
cryptography>=41.0.3

# AI & ML
openai>=1.3.0

# Basic ML (lighter alternatives)
scikit-learn>=1.3.0
numpy>=1.24.3
pandas>=2.0.3

# GUI & Visualization
matplotlib>=3.7.2

# Web Dashboard
flask>=3.0.0
flask-socketio>=5.3.0

# System & Process Monitoring
psutil>=5.9.5
pyobjc-framework-Cocoa; sys_platform == "darwin"

# Utilities & Logging
python-dotenv>=1.0.0
loguru>=0.7.0
click>=8.1.7
schedule>=1.2.0
pydantic>=2.3.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0 